real_time_pipeline/publisher/requirements.txt
# real_time_pipeline/publisher/requirements.txt
google-cloud-pubsub==2.13.0
pandas==1.5.3
numpy==1.23.5